convert access by value to access by reference
